OBG-904. Refactored dependencies
